donation_mgt_sys/requirements.txt
asgiref==3.7.2
crispy-bootstrap5==2023.10
Django==5.0.1
django-crispy-forms==2.1
#mysqlclient==2.2.1
pillow==10.2.0
sqlparse==0.4.4
typing_extensions==4.9.0
